requirements.txt
colorlog==5.0.1
python-magic==0.4.22
pytest==5.4.1
